# backend/docs/api-specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RoomMate Expense Management API",
    "description": "API cho ứng dụng quản lý chi tiêu chung giữa các roommate",
    "version": "1.0.0",
    "contact": {
      "name": "Development Team",
      "email": "dev@roommate-expense.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Expense": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "title": {
            "type": "string",
            "example": "Grocery Shopping"
          },
          "amount": {
            "type": "string",
            "example": "150.75"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Weekly grocery shopping at supermarket"
          },
          "payerId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "isShared": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/receipt.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Settlement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "payerId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "payeeId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "amount": {
            "type": "string",
            "example": "75.50"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Settlement for grocery expenses"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/transfer-receipt.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateExpenseRequest": {
        "type": "object",
        "required": ["title", "amount"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Grocery Shopping"
          },
          "amount": {
            "type": "string",
            "example": "150.75"
          },
          "description": {
            "type": "string",
            "example": "Weekly grocery shopping"
          },
          "isShared": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/receipt.jpg"
          }
        }
      },
      "CreateSettlementRequest": {
        "type": "object",
        "required": ["payeeId", "amount"],
        "properties": {
          "payeeId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "amount": {
            "type": "string",
            "example": "75.50"
          },
          "description": {
            "type": "string",
            "example": "Settlement for shared expenses"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/transfer-receipt.jpg"
          }
        }
      },
...
      "ExpenseResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Expense created successfully"
          },
          "expense": {
            "$ref": "#/components/schemas/Expense"
          }
        }
      },
      "ExpensesListResponse": {
        "type": "object",
        "properties": {
          "expenses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          }
        }
      },
      "SettlementResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Settlement created successfully"
          },
          "settlement": {
            "$ref": "#/components/schemas/Settlement"
          }
        }
      },
      "SettlementsListResponse": {
        "type": "object",
        "properties": {
          "settlements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Settlement"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid input data"
          },
          "message": {
            "type": "string",
            "example": "Detailed error message"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/expenses": {
      "get": {
        "summary": "Get All Expenses",
        "description": "Lấy danh sách tất cả chi tiêu",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách chi tiêu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpensesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create New Expense",
        "description": "Tạo khoản chi tiêu mới",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo chi tiêu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{id}": {
      "put": {
        "summary": "Update Expense",
        "description": "Cập nhật thông tin chi tiêu",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID của khoản chi tiêu"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền cập nhật",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy chi tiêu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Expense",
        "description": "Xóa khoản chi tiêu",
        "tags": ["Expenses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID của khoản chi tiêu"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Expense deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền xóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy chi tiêu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settlements": {
      "get": {
        "summary": "Get All Settlements",
        "description": "Lấy danh sách tất cả thanh toán",
        "tags": ["Settlements"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách thanh toán",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create New Settlement",
        "description": "Tạo thanh toán mới",
        "tags": ["Settlements"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSettlementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thanh toán thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}